20241107 1751

+++ src/vpa_robot_vision/scripts/hsv/search_pattern.py
@@ -192,8 +192,8 @@ def search_inter_guide_line2(hsv_space:HSVSpace, hsv_image, action:int):
         res = None
         for i in range(60,170,20):
             y = height - i
-            line1 = np.nonzero(mask[y, : width / 2])[0]
-            line2 = np.nonzero(mask[y-20, : width / 2])[0]
+            line1 = np.nonzero(mask[y, : int(width / 2)])[0]
+            line2 = np.nonzero(mask[y-20, : int(width / 2)])[0]
             # print(y,line)
             seg1 = _break_segs(line1)
             seg2 = _break_segs(line2)
